REFACTOR: updated file names
