+++ __tests__/Unit/Componets/Button/Button.test.tsx
@@ -1,4 +1,4 @@
-import Button from "@/components/Button/Button";
+import Button from "@/components/Button";
 import { render, screen, fireEvent } from "@testing-library/react";
 
 describe("Button Test", function () {
