+++ src/components/VotesListModal/VotesDescription.tsx
@@ -1,4 +1,4 @@
-import Avatar from "../Avatar/Avatar";
+import Avatar from "../Avatar";
 
 type VotesTypes = {
     src?: string;
